+++ src/oemof/network/network/nodes.py
@@ -67,14 +67,14 @@ class Node(Entity):
         self._outputs = Outputs(self)
         self._in_edges = set()
 
+        self._parent = None
+        self._depth = 0
+
+        self._subnodes = []
+        self._energy_system = None
+
         if parent_node is not None:
             parent_node.add(self)
-        else:
-            self._parent = None
-            self._depth = 0
-
-        self.__subnodes = []
-        self.__energy_system = None
 
         # TODO: Try to avoid this local `import`.
         from ..energy_system import EnergySystem
@@ -147,7 +147,7 @@ class Node(Entity):
         It is deliberately provided as a tuple to prevent user to append
         subnodes other than with API methods.
         """
-        return tuple([sn for sn in self.__subnodes])
+        return tuple([sn for sn in self._subnodes])
 
     @property
     def parent(self) -> Node:
@@ -159,9 +159,10 @@ class Node(Entity):
         for subnode in subnodes:
             subnode._parent = self
             subnode._depth = self.depth + 1
-            self.__subnodes.append(subnode)
-            if self.__energy_system is not None:
-                self.__energy_system.add(subnode)
+            self._subnodes.append(subnode)
+            if self._energy_system is not None:
+                self._energy_system.add(subnode)
+                subnode._energy_system = self._energy_system
 
     def subnode(self, class_, local_name, *args, **kwargs):
         """Create a subnode and add it to this `Node`.
@@ -229,9 +230,9 @@ class Node(Entity):
         #    Explain why the `node` argument is necessary.
         if self is not node:
             raise ValueError("Call needs to be obj._add_subnodes(obj).")
-        self.__energy_system = kwargs["EnergySystem"]
+        self._energy_system = kwargs["EnergySystem"]
         deque(
-            (kwargs["EnergySystem"].add(sn) for sn in self.__subnodes),
+            (kwargs["EnergySystem"].add(sn) for sn in self._subnodes),
             maxlen=0,
         )