+++ CHANGELOG.rst
@@ -48,6 +48,7 @@ Changelog
 * Properties "results", "timeindex", "timeincrement", and "temporal"
   of the class EnergySystem are now deprecated. They are not used (at least
   not at the level of oemof.network).
+* Fixed passing references to the EnergySystem when adding Nodes as subnodes.
 * The required python version is upped to 3.10. This prevents problems in
   environment creation and moves with the end-of-life-timeline of python 
   versions.